# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: BD E-Commerce Platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /categories:
    get:
      operationId: listCategories
      tags: [categories]
      summary: List all categories
      responses:
        "200":
          description: Category list
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/Category" }
    post:
      operationId: createCategory
      tags: [categories]
      summary: Create a category
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CategoryInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Category"
  /categories/{id}:
    patch:
      operationId: updateCategory
      tags: [categories]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CategoryInput"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Category"
    delete:
      operationId: deleteCategory
      tags: [categories]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      responses:
        "204":
          description: Deleted

  # ── BRANDS ──────────────────────────────────────────────────
...
  /banners:
    get:
      operationId: listBanners
      tags: [banners]
      summary: List homepage banners
      responses:
        "200":
          description: Banner list
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/Banner" }
    post:
      operationId: createBanner
      tags: [banners]
      summary: Create a banner
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BannerInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Banner"
  /banners/{id}:
    delete:
      operationId: deleteBanner
      tags: [banners]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      responses:
        "204":
          description: Deleted

  # ── SITE SETTINGS ──────────────────────────────────────────────
...
  /coupons/validate:
    post:
      operationId: validateCoupon
      tags: [coupons]
      summary: Validate a coupon code
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CouponValidate"
      responses:
        "200":
          description: Coupon info
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Coupon"
        "400":
          description: Invalid or expired coupon
  /coupons:
    get:
      operationId: listCoupons
      tags: [coupons]
      summary: List all coupons (admin)
      responses:
        "200":
          description: Coupon list
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/Coupon" }
    post:
      operationId: createCoupon
      tags: [coupons]
      summary: Create a coupon (admin)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CouponInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Coupon" 

  # ── DASHBOARD ──────────────────────────────────────────────
...
components:
  schemas:
...
    Category:
      type: object
      required: [id, name, slug]
      properties:
        id: { type: integer }
        name: { type: string }
        slug: { type: string }
        image: { type: ["string", "null"] }
        productCount: { type: integer }
    CategoryInput:
      type: object
      required: [name, slug]
      properties:
        name: { type: string }
        slug: { type: string }
        image: { type: ["string", "null"] }

    # ── Brand ──────────────────────────────────────────────────
...
    Banner:
      type: object
      required: [id, imageUrl]
      properties:
        id: { type: integer }
        imageUrl: { type: string }
        title: { type: ["string", "null"] }
        link: { type: ["string", "null"] }
        active: { type: boolean }
    BannerInput:
      type: object
      required: [imageUrl]
      properties:
        imageUrl: { type: string }
        title: { type: ["string", "null"] }
        link: { type: ["string", "null"] }
        active: { type: boolean }

    # ── Settings ──────────────────────────────────────────────────
...
    Coupon:
      type: object
      required: [id, code, discountType, discountAmount]
      properties:
        id: { type: integer }
        code: { type: string }
        discountType: { type: string, enum: [percentage, fixed] }
        discountAmount: { type: number }
        minOrderAmount: { type: ["number", "null"] }
        expiryDate: { type: ["string", "null"] }
        active: { type: boolean }
    CouponInput:
      type: object
      required: [code, discountType, discountAmount]
      properties:
        code: { type: string }
        discountType: { type: string, enum: [percentage, fixed] }
        discountAmount: { type: number }
        minOrderAmount: { type: ["number", "null"] }
        expiryDate: { type: ["string", "null"] }
        active: { type: boolean }
    CouponValidate:
      type: object
      required: [code]
      properties:
        code: { type: string }
        orderAmount: { type: number }

    # ── Dashboard ──────────────────────────────────────────────────
...