# lib/api-zod/src/generated/types/couponValidate.ts
/**
 * Generated by orval v8.22.0 🍺
 * Do not edit manually.
 * Api
 * BD E-Commerce Platform API
 * OpenAPI spec version: 0.1.0
 */

export interface CouponValidate {
  code: string;
  orderAmount?: number;
}